Enemy Patrol State
There are now 3 enemies that wander their own specified paths in a loop, and when the player comes into contact with them the game stops. I can easily add more enemies and paths for them to follow. Next up will be creating more states such as going idle occasionally and chasing the player.